Add files via upload
Submission contains all three types of CSS implementation, headers, paragraphs, a link, an image, an unordered list, a table, and a form. I don't think there's much to reflect on at this point in time, as it was not part of the assignment to make sure the site is coherent or actually looks good, just to demonstrate the capacity to use the requested elements.